+++ python/bulkpublish/client.py
@@ -43,6 +43,7 @@ from .exceptions import (
 from .labels import AsyncLabelsResource, LabelsResource
 from .hashtag_groups import AsyncHashtagGroupsResource, HashtagGroupsResource
 from .templates import AsyncTemplatesResource, TemplatesResource
+from .review_links import AsyncReviewLinksResource, ReviewLinksResource
 from .calendar_notes import AsyncCalendarNotesResource, CalendarNotesResource
 from .media import AsyncMediaResource, MediaResource
 from .posts import AsyncPostsResource, PostsResource
@@ -175,6 +176,7 @@ class BulkPublish(_BaseClient):
         labels: :class:`~bulkpublish.labels.LabelsResource`
         hashtag_groups: :class:`~bulkpublish.hashtag_groups.HashtagGroupsResource`
         templates: :class:`~bulkpublish.templates.TemplatesResource`
+        review_links: :class:`~bulkpublish.review_links.ReviewLinksResource`
         calendar_notes: :class:`~bulkpublish.calendar_notes.CalendarNotesResource`
         schedules: :class:`~bulkpublish.schedules.SchedulesResource`
         channel_sets: :class:`~bulkpublish.channel_sets.ChannelSetsResource`
@@ -234,6 +236,7 @@ class BulkPublish(_BaseClient):
         self.labels = LabelsResource(self)
         self.hashtag_groups = HashtagGroupsResource(self)
         self.templates = TemplatesResource(self)
+        self.review_links = ReviewLinksResource(self)
         self.calendar_notes = CalendarNotesResource(self)
         self.schedules = SchedulesResource(self)
         self.channel_sets = ChannelSetsResource(self)
@@ -554,6 +557,7 @@ class AsyncBulkPublish(_BaseClient):
         self.labels = AsyncLabelsResource(self)
         self.hashtag_groups = AsyncHashtagGroupsResource(self)
         self.templates = AsyncTemplatesResource(self)
+        self.review_links = AsyncReviewLinksResource(self)
         self.calendar_notes = AsyncCalendarNotesResource(self)
         self.schedules = AsyncSchedulesResource(self)
         self.channel_sets = AsyncChannelSetsResource(self)